Consistently use 64-bit integers in `io.ascii` (including Windows) (#16005)
* ensure 64 bit int columns remain 64 bit int type on Windows (Python path)

* ensure 64 bit in columns read correctly, C code path

* Fixed tests to reflect changes from reading 64bit integers as ints

* C path int parsing: use int64_t instead of long long

* update the test to reflect the policy of defaulting to 64bit ints in all cases.

* update changelog to stress  the 64-bit by default behavior.

* forward port ipac test from #15992

* small optimization: inline int64 string parsing C function

* Tweak changelog wording and add what's new entry

+++ astropy/io/ascii/tests/test_c_reader.py
@@ -1442,8 +1442,8 @@ def test_int_out_of_range(parallel, guess):
     Integer numbers outside int range shall be returned as string columns
     consistent with the standard (Python) parser (no 'upcasting' to float).
     """
-    imin = np.iinfo(int).min + 1
-    imax = np.iinfo(int).max - 1
+    imin = np.iinfo(np.int64).min + 1
+    imax = np.iinfo(np.int64).max - 1
     huge = f"{imax+2:d}"
 
     text = f"P M S\n {imax:d} {imin:d} {huge:s}"
@@ -1484,7 +1484,7 @@ def test_int_out_of_order(guess):
     shows up first, it will produce a string column - with both readers.
     Broken with the parallel fast_reader.
     """
-    imax = np.iinfo(int).max - 1
+    imax = np.iinfo(np.int64).max - 1
     text = f"A B\n 12.3 {imax:d}0\n {imax:d}0 45.6e7"
     expected = Table([[12.3, 10.0 * imax], [f"{imax:d}0", "45.6e7"]], names=("A", "B"))
